+++ backend/src/main/java/com/daily/daily/member/dto/PasswordRecoverDTO.java
@@ -1,0 +1,4 @@
+package com.daily.daily.member.dto;
+
+public class PasswordRecoverDTO {
+}
